CPURender类绘制字体采用GDI绘制 大幅度提升绘制速度

+++ examples/KuGou/LrcControl.h
@@ -16,7 +16,7 @@ public:
 class LrcControl :
 	public Control
 {
-	int offsetY = 0;
+	float offsetY = 0;
 	Lrc* LrcNow = NULL;
 	/// <summary>
 	 /// 上下居中的Y坐标值
